Build x86_64-pc-windows-gnu target with the msvc toolchain
This makes it a cross-build like i686-pc-windows-msvc and
aarch64-pc-windows-msvc already are.

The situation is not truly symmetric, and there is probably no
argument from elegance for doing this. Instead, it is in the hope
that it will be easier to install the tools we need without
compiling them, thereby saving time and making it reasonable to use
more tools, including those needed for auditable builds.