Say why we test gix-hashtable instead of gix on armv7
This completes the TODO about investigating what was thought to be
an environment error, possibly because it references `LD_PRELOAD`.
We can't run `cross test -p gix --target armv7-linux-androideabi`
because the armv7 `bash` that runs the test fixture scripts (whose
archives are `.gitignore`d) finds the `git` command on the runner,
which is an amd64 build, and tries to run it as though it is a
dynamically linked armv7 executable. This relates to the use of
`cross`; no analogous problem occurs with the i686 builds, which
are able to run the amd64 `git` avaialble on the system.

There are various ways to try to fix this, but for now we do not
have any such fix and this is putting it back to testing with
`-p gix-hashtable` on armv7 (while still using `-p gix` on i686),
and also replacing the old TODO comment with comments on the newly
introduced matrix variable about why `gix` is tested on one target
and `gix-hashtable` is tested on another.