Install 64-bit Node in container, for its libraries
This is a workaround for a problem that is the same or similar to
https://github.com/actions/checkout/issues/334:

    Run actions/checkout@v4
    /usr/bin/docker exec  a7f94cb30b99bd1ee851270fad4768421eddecd7e8fa674f56c4f2d2d64210b6 sh -c "cat /etc/*release | grep ^ID"
    exec /__e/node20/bin/node: no such file or directory

I had not originally noticed the `cat /etc/*release` part of that
output, and had surmised that this was a subtly different problem,
but it looks like it's probably exactly the same. The workaround in
https://github.com/actions/checkout/issues/334#issuecomment-2435101135
is applied in this commit and seems to be working.

As noted there, the absence of some 64-bit libraries in 32-bit
containers prevents the (existing and already mapped) 64-bit
Node.js outside the container from being run, and installing 64-bit
Node.js in the container via `apt-get` (even though, in this case,
it is a different version) installs the needed dependencies,
letting the (existing) Node.js run.

(Other problems, related to `safe.directory` protections as well as
tests that are known to fail due to assertions about the size of
data structures that implicitly assume 64-bit targets, do still
prevent the job from succeeding. But `actions/checkout@v4`, which
was failing before, seems to be working, as do the other actions.)