Revert "Try not caching journey test dependencies"
This is part of investigating the problem of not being able to
connect to the git daemon. This problem is described in:
https://github.com/GitoxideLabs/gitoxide/pull/1634#pullrequestreview-2507044578

The preceding commit, bf2abdb41e22e27f599e845152b0b4643ec17235,
which this reverts, omitted caching from the `test-journey` CI job,
and that somehow avoided the problem, causing all tests to pass.
With that change, the failure did not recur even after rerunning
the job more than 15 times:
https://github.com/EliahKagan/gitoxide/actions/runs/12363538568

That is strange, because no cause related to Rust dependencies is
apparent: The Git daemon is not itself provided by or expected to
be affected by them, and any stale dependencies from the cache
should automatically not be used.

However, perhaps the extra *delays*, incurred by not having cached
some artifacts and having to rebuild them, allowed the
failure-triggering condition to resolve on its own. If so, that
would explain this observation.

It would also be consistent with how the problem arose with the
changes in #1634, if a previous delay existed that happened to
prevent the bug from being observed and that #1634 somehow
counteracted. This could be either by a performance improvement,
or by shifting the delays to other times during the job (by
changing which artifacts were needed for which tests).

The `test-journey` job usually takes between 6:00 an 6:30 without
caching build artifacts. This is arguably not too bad, so it might
be feasible to use that change as a long-term workaround. But if
possible the underlying problem should be identified and fixed.

So this commit brings back caching to check if it reliably restores
the bug. If so, then experiments and other fixes or workarounds can
be attempted on top of that.