Use `String` paths and replace `\` with `/`, in new tests
This fixes the two failing `*_indirect` tests on Windows. The
three failing `*_direct` tests still fail. This is for a different
reason related to the meaning of paths like `/bin/sh` on Windows,
which would need to be interpreted differently form their
drive-relative meaning as Windows paths, in order to work correctly
in `#!` lines.