Try to enable ports APT sources unconditionally
This does not work as written, at least on s390x, because the ports
list is already written there from the base image. The existing
file is actually better than what this writes, though what this
writes is simpler and so might potentially be easier to debug or
reason about. Unfortunately, if we clobber the existing file with
this one, the error installing `git:s390x` is unchanged.

Uninstalling the old `amd64` `git` package first (not shown here)
and packages installed as its dependencies doesn't help either.