[Test commit] Tell Dependabot not to use conventional commits
Dependabot automatically examines existing commits' messages to
decide what style to use, if no `commit.prefix` is configured. But
gitoxide prefers that conventional commits be used only in limited
circumstances: the "purposeful conventional commits" style detailed
in `DEVELOPMENT.md`. Commits to bump dependencies should typically
not be conventional, so Dependabot should not title its commits
with prefixes that can be interpreted as meaningful conventional
commits. But the style Dependabot has automatically picked here
uses prefixes such as `build(deps): ...`.

In practice this does not cause serious problems, and it would be
fine to use Dependabot even if this behavior couldn't be changed.
But there are a few reasons to configure Dependabot not to do this:

- Stylistic consistency.

- Decrease the likelihood of future undesired behavior if a future
  version of `cargo-smart-release` treats more conventional commit
  prefixes specially.

- Make commit messages slightly easier to read if one is familiar
  with this repository.

- Avoid giving newcomers the wrong idea about how, and how often,
  commit messages in this project are expected to be conventional.

This attempts to configure Dependabot not to use any prefix.