Revert "Try specifying prefix as `null` rather than `''`"
This reverts commit e53c2dc4701e55a8ed24211a90b56ce77d4ccbc2, which
failed (https://github.com/EliahKagan/gitoxide/runs/40278979940):

    Dependabot encountered the following error when parsing your `.github/dependabot.yml`:

        The property '#/updates/0/commit-message/prefix' of type null did not match the following type: string

    Please update the config file to conform with Dependabot's specification.

This is to say that while an empty prefix can be used, a value of
`null` does not enable the behavior of having no prefix, as hoped.

The distinction between an empty prefix and no prefix is that, with
no prefix, Dependabot capitalizes the title of the commit message.
However, without specifying a prefix, i.e. when the key is absent
from the configuration, Dependabot automatically detects whether to
use a prefix and what prefix to use based on preexisting commit
messages. So it seems there is no way to force it to use no prefix.