Install missing `rustup` in `windows-11-arm` jobs
This works around the combination of these two conditions, which
individually wouldn't be a problem but are a problem together:

- The newly available `windows-11-arm` runner does not currently
  have `rustup` installed on it.

- Although the `dtolnay/rust-toolchain` action will attempt to
  install `rustup` when it is absent on Unix-like systems, it does
  not attempt to do so on Windows.

(If the latter condition is fixed, however, then depending on *how*
it is fixed, it might still be necessary to install `rustup`
ourselves. Specifically, it is easy to accidentally download a
`rustup-init.exe` for x86-64 even when on ARM64/AArch64 Windows.)

The new step here, added for both of the `windows-11-arm` jobs, is
intended to be similar in its effect to these commands that
`dtolnay/rustup-init` runs on Unix-like systems:
https://github.com/actions-rust-lang/setup-rust-toolchain/blob/9399c7bb15d4c7d47b27263d024f0a4978346ba4/action.yml#L136-L139

Note that this is not quite equivalent. It is intentionally less
versatile in two ways:

- It hard-codes the AArch64 `rustup-init.exe` URL, to ensure that
  it is selected, rather than not finding anything or wrongly
  getting a URL for a Unix-like system or for x86-64 Windows.

- No attempt is made to respect a preexisting `CARGO_HOME`
  environment variable (since we do not set one in these jobs).

(This uses the style `$Env:varname` rather than `$env:varname` in
PowerShell because the former seems to be much more common in the
Microsoft documentation, and changes one preexisting occurrence
of `$env` to `$Env` for stylistic consistency.)