Use tmpfs-backed ext4 as a workaround
For when tmpfs does not suppport `-o casefold` in mounting.

This does require that ext4 support `-O casefold`, but that was
already working before.

Even though that worked before, this could fail when using an ext4
image on tmpfs, because the ext4 image is smaller than it was when
tmpfs was not used, in order to allow it to be created in memory.
While this allows it to be created, the tests may fail if build
artifacts and other files need more space than is available.