Run only the test(s) of interest
In `test-linux-casefold`, this runs only the tests that are
duplicates of each other, where the failure has been observed,
rather than first running the whole test suite.

The reason is to try to build fewer crates and test executables,
since this is currenntly failing because it runs out of space on
the ext4 image whose size is itself constrained by the amount of
available memory for the tmpfs filesystem on which the image is
created.