Try to fail more often, with less delay
The `test-ext4-casefold` jobs were still not surfacing the error
quite reliably enough. The wait was also longer than ideal. This
commit makes some changes that hope to improve on both.

- Decrease the number of duplicate test cases by a factor of 4.
  This has the effect that they are increased by 25x compared to
  the multiplier in the commited code, instead of 100x.

- Create 10 times as many CI jobs in the matrix, by adding a `num`
  variable whose values are those in the range 0..=9. This variable
  is not actually used, it just serves to produce more CI jobs.

- Don't use `Swatinem/rust-cache` in these jobs anymore. It is not
  obvious that it will make things work better overall, in terms of
  speedup, cache usage, and cache access, now that we are
  increasing the number of matrix jobs by a factor of 10.

Moving some of the repetition into separate jobs, which may run in
parallel, is not only, nor even primarily, to leverage runner
parallelism to do more test runs faster. Instead, it is because it
looks like some chaotic initial state may contribute to how likely
the failure is to occur. No definitive experiment on CI shows this
to be the case -- but in local testing, on some systems, I often
have runs that fail early over and over again, and then wait a
while, come back, and have many runs that don't surface the bug,
then come back later and it is easy to produce again, and so on.