Undiversify runners but vary test parallelism
Since varying `runs-on` didn't surface more errors, nor help smooth
out inconsistencies across runs.

The `*-arm` runners seem to have fewer errors with the particular
way the tests are running now, so this just uses `ubuntu-latest`.

This also adjusts `rep`, having it take on more values, so as to
more than make up for what would otherwise be a smaller number of
jobs.