Test more values for test parallelism
This creates additional CI `test-ext4-casefold` jobs, for more
values of the `--test-threads` operand passed to `cargo nextest`.

The motivation is that the value of 3 curiously seems never to
surface the failure on CI. Unlike 1, 3 should is expected to
surface the failure comparably to 2 and 4, both of which do fail
more often than not. Yet the jobs with `--test-threads=3` keep
passing.

(This is unlikely to be due to case-sensitivity misdetection, since
the assertions for when the filesystem is case-sensitive would also
fail, if they fire in a job where the filesystem is case-folding.)

If these jobs are kept, then `rep` should probably be adjusted to
have fewer, perhaps 7, values. However, this adjusts `rep` to have
16 values (i.e. twice as many as before), since the higher operands
to `--test-threads` are expected to produce the failure less often.