Test fewer test parallelism values and reps
- Test `--test-threads` operands of 1 to 4, rather than 1 to 7.
- Do 7 reps instead of 16.

In addition to the 1000x repetition in `checkout.rs` of the
writes_through_symlinks_are_prevented_even_if_overwriting_is_allowed
test case -- which almost always surfaces #2006 on CI on macOS and
Windows without further modifications -- the `test-ext4-casefold`
jobs as they currenly stand seem like they would ber sufficient to
test whether a possible future solution would fix the bug on
GNU/Linux.

This looks like about the minmum amount we should reasonably verify
still fails (in some jobs) immediately before a #2006 fix, while no
longer failing *any* jobs afterwards. But the `test-ext4-casefold`
jobs are still much more than would be wanted on the main branch.

Assuming the cause of the failures is the same on all systems, and
the solution neither explicitly nor implicitly operates differently
across systems (when comparing how it works when the filesystem is
case-insensitive, that is), it may be enough on CI to regression
test only macOS and Windows for #2006.

Thus, this also adds a TODO comment atop the `test-ext4-casefold`
definition reminding that it should not be kept, or not in full.