Move published release `installation` test to new workflow
So that it can be triggered from the `release` workflow (as they
have been since #1555), as well as directly.

This is to make it easier to run or rerun the published crate
installation tests, as well as any future such tests that may be
added, without having to run jobs that are involved in making the
release.