Skip performance tests on Windows 11 ARM64 CI
This is to investigate whether fixtures for other tests are failing
due to interaction with them, and it is hoped to be a temporary
measure. Temporary or not, this change affects only Windows ARM
jobs. Other ARM platforms, and non-ARM Windows, are unchanged.

This might remain in place over a longer term:

- If the problem is that the current Windows 11 ARM runner is not
  fast enough to allow the tests to pass. Though adjusting expected
  timings conditionally for Windows ARM64 might be a workaround.

- If including the performance tests, passing are not, makes the
  runs take too long to do routinely. It looks like this may be the
  case. It's probably better to test regularly on Windows ARM
  without the performance tests, than to rarely test Windows ARM.