Show details on ARM Git for Windows before and after upgrade
The version of Git for Windows after the upgrade (or without the
upgrade, if it is not done) is already shown in `actions/checkout`
output, since that action defaults to using the installed Git.

However, it's valuable to see information about it before and after
upgrade in the same run. It's also useful to see more details than
the version, such as to confirm the assumption that it is really
an ARM64 build.

All the information of interest could be obtained by running
`git version --build-options`. But the change here continues to
avoid running the old Git for any reason. This instead downlods
and installs `sigcheck` to inspect a (non-shim) `git.exe`.

This is with the idea that running any executables provided with
Git for Windows may increase the likelihood of being unable to
replace them, if somehow they or their subprocesses do not fully
terminate.

This is probably the wrong tradeoff, and it will probably be
changed to running `git version --build-options` before and after
the upgrade. But if that forthcoming change does cause a problem,
then it can be reverted and the approach done here used again.