Get details on ARM Git for Windows on CI from Git itself
This replaces the more complicated `sigcheck` logic with runs of
`git version --build-options` before and after upgrading Git for
Windows in the ARM `test-fixtures-windows` job.

The reason this wasn't done before was to avoid running the old
version at all, to maximize the chance that Inno Setup would be
able to remove it. On Windows, it is often difficult or infeasible
to remove or replace a running executable or other open file,
without terminating the executable or closing the file.

However, this was not the best tradeoff. Running `git version` is
unlikely to leave a danling process running. What we really want is
to avoid using the old Git substantially, such as in
`actions/checkout`, and this still avoids that.

The other reason to upgrade Git before running `actions/checkout`
is to verify that the new version of Git for Windows is working by
having `actions/checkout` use it. This change also preserves that.

Passing `--build-options` is important here because we do want to
verify that both the version of Git for Windows being replaced, and
(especially) the upgrade, are ARM64 builds.

(The `windows-11-arm` image currently has x86-64 builds of some
preinstalled software. But it's not expected to have a non-native
build of Git for Windows. If it does, then there may be more steps
that should be taken besides attempting to upgrade it.)