Try hard-coding the repository
Though I don't expect that to work, since I think the `github`
context is just not available in `uses`.

Also, if it does work, I'm not sure how robust it would be. In
common scenarios, a commit is available by its OID througout the
fork network. However, when this is a merge commit for an unmerged
PR (which is what `github.sha` is in workflow runs triggered by the
`pull_request` event), I am uncertain if it is actually guaranteed
to be available when accessed on repositories in the fork network
*other* than the repository from which the PR originates and the
repository the PR is opened on.