Try `jenseng/dynamic-uses`
For the call to the extracted `upgrade-git-for-windows` composite
action.

Though even if this works, the conceptual heft (complexity of what
we are doing and why, even though the code is simple), together
with making it harder for automated tools to know it's a composite
action in our that is being indirectly used, may make this approach
unworthwhile compared to just keeping the steps inlined in
`ci.yml`.