Temporarily break `wasm` feature of `gix-diff` to check CI
As noted in #2092, the `wasm` jobs on CI do not test `gix-diff`
directly. However, my prediction there that breakage would not be
detected on CI is likely mistaken, because those jobs do test the
`wasm` feature of `gix-pack`. The `gix-pack` crate depends on
`gix-diff`, and its `wasm` feature enable the `gix-diff` one.

This temporary change checks that CI does fail when the `wasm`
feature of `gix-diff` is broken in a simple way, even though it
does not currently run WASM tests of `gix-diff` directly.

This also temporarily makes the `wasm` matrix non-fail-fast, so
that more failures can be observed.