Prevent unintended expansions in `make_submodules` fixture script
This quotes the command substitution so that if the current working
directory path contains IFS whitespace such as spaces, or contains
globbing characters, these will be treated literally, rather than
triggering word splitting or pathname expansion.

(Much less importantly, this also changes it to use the $() syntax
for command substitution. This is separate from quoting, since both
syntaxes need to be double-quoted to avoid splitting and globbing.)