Test `winget` as another way to upgrade GfW on CI
Upgrading Git for Windows is not usually needed at all. The idea
here is to test `winget` to see if it can do it more simply than
the code that has been added and removed multiple times before
that downloads the installer from the GitHub release and runs it,
passing a number of non-obvious InnoSetup arguments.

If this works, then it could be a one-liner, but we would still not
want to do it by default, and also it may be that not all runners
will have `winget` installed or that not all runners will have
`winget` set up in a usable way.

This applies only to Windows, of course.