Abandon `winget` idea; start on upgrading GfW the old way
So far this only attempts to upgrade it for `test-fast` because
whatever is going to go wrong may as well be fixed for that before
extending it to `test-fixtures-windows`.

The approach followed here is similar to what was done in the past,
but some specific details differ, including trying to make it work
for both ARM64 and x64 systems and putting everything in a single
step rather than three steps.