Upgrade GfW in `test-fixtures-windows` too, without duplication
GitHub Actions has recently started supporting YAML anchors:

- https://github.blog/changelog/2025-09-18-actions-yaml-anchors-and-non-public-workflow-templates/
- https://docs.github.com/en/actions/reference/workflows-and-actions/reusing-workflow-configurations#yaml-anchors-and-aliases

This uses that to have the same step in two separate job
definitions without duplicating the code itself.

It would often be preferable do put the job in a composite action,
but as noted in previous commit messages (especially in the
parenthesized paragraph at the end of 9e4e3ec), this job is best
run before `actions/checkout`, yet retrieving the action definition
from the current repository at the same ref (in a way that is not
excessively complicated or slow) requires that `actions/checkout`
have been used first. So this uses a YAML anchor instead.